ai_module/requirements.txt
# Core RAG AI Dependencies for Ticketportaal
# Python 3.11+

# Vector Database
chromadb==0.4.15

# Embedding Model
sentence-transformers>=2.7.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Additional Dependencies
pydantic==2.4.2
requests==2.31.0
python-multipart==0.0.6

# Database Connectivity
mysql-connector-python==8.2.0

# NER and Entity Extraction
spacy>=3.7.0

# Knowledge Graph
networkx>=3.2.0

# BM25 Sparse Keyword Search
rank-bm25>=0.2.2

# Utilities
python-dotenv==1.0.0
